Setup DDD structure. Move files and modules to appropriate locations
