+++ src/application/app.module.ts
@@ -1,8 +1,6 @@
 import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
-import { ConfigModule } from './config/config.module';
-import { ConfigService } from './config/config.service';
+import { ConfigModule } from '../infrastructure/config/config.module';
+import { ConfigService } from '../infrastructure/config/config.service';
 import { MongooseModule } from '@nestjs/mongoose';
 
 @Module({
@@ -10,7 +8,7 @@ import { MongooseModule } from '@nestjs/mongoose';
     inject: [ConfigService],
     useFactory: async (configService: ConfigService) => configService.getMongoConfig(),
   }),],
-  controllers: [AppController],
-  providers: [AppService],
+  controllers: [],
+  providers: [],
 })
 export class AppModule { }
